@@ -7,7 +7,7 @@
   "bugs": {
     "url": "http://github.com/EXXETA/kubernetes-monitor-view/issues"
   },
-  "version": "0.1.12",
+  "version": "0.2.0",
   "license": "Apache License 2.0",
   "scripts": {
     "ng": "ng",
@@ -22,57 +22,59 @@
   },
   "private": true,
   "peerDependencies": {
-    "@angular/core": ">=^5.0.0"
+    "@angular/core": ">=^6.0.0"
   },
   "dependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.8.0",
-    "@angular-devkit/core": "^0.8.2",
-    "@angular-devkit/schematics": "^0.8.2",
-    "@angular/animations": "^6.1.0",
-    "@angular/cli": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular-devkit/build-angular": "^0.13.9",
+    "@angular-devkit/build-ng-packagr": "~0.8.0",
+    "@angular-devkit/core": "^0.8.9",
+    "@angular-devkit/schematics": "^0.8.9",
+    "@angular/animations": "^6.1.10",
+    "@angular/cli": "^6.2.9",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/language-service": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/router": "^6.1.10",
     "@ng-bootstrap/ng-bootstrap": "^2.0.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "bootstrap": "^3.3.7",
-    "bootstrap-sass": "^3.3.7",
-    "codelyzer": "^4.0.1",
-    "core-js": "^2.4.1",
-    "css-loader": "^0.28.11",
+    "@types/bootstrap": "^4.3.0",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^12.0.4",
+    "bootstrap": "^4.3.1",
+    "codelyzer": "^4.5.0",
+    "core-js": "^2.6.9",
+    "css-loader": "^2.1.1",
     "extract-text-webpack-plugin": "^3.0.2",
     "fs-extra": "^5.0.0",
-    "jquery": "3.2.1",
+    "jquery": "^3.4.1",
     "jquery-ui-dist": "^1.12.1",
-    "karma": "^3.1.1",
+    "karma": "^4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~1.4.2",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "moment": "^2.22.2",
-    "ng-packagr": "^4.1.1",
-    "ngx-logger": "^3.3.4",
+    "moment": "^2.24.0",
+    "ng-packagr": "^5.2.0",
+    "ngx-logger": "^3.3.13",
     "ngx-toastr": "^8.6.0",
-    "node-sass": "^4.9.0",
-    "octicons": "^8.1.0",
-    "popper.js": "^1.14.3",
-    "protractor": "^5.4.1",
-    "rxjs": "^6.2.2",
+    "octicons": "^8.5.0",
+    "popper.js": "^1.15.0",
+    "protractor": "^5.4.2",
+    "rxjs": "^6.5.2",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.21.0",
     "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.7.2",
-    "zone.js": "^0.8.19"
+    "tsickle": ">=0.29.0",
+    "tslib": "^1.9.0",
+    "tslint": "^5.17.0",
+    "typescript": "^3.4.5",
+    "zone.js": "^0.8.29"
   }
 }
