@@ -24,10 +24,9 @@ import { KubernetesMonitorMockService } from './kubernetesMonitorMock.service';
 
 
 
-
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
   ],
   imports: [
     BrowserModule,
