@@ -15,37 +15,49 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { StatusReport } from './model/StatusReport';
+import { map } from 'rxjs/operators';
+import { Domain } from './model/Domain';
+import { HttpClient } from '@angular/common/http';
+import {Observable} from 'rxjs/internal/Observable';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export abstract class KubernetesMonitorService {
 
-
+  constructor(private http: HttpClient) {
+  }
 
   public selectedApplicationName: string;
   public selectedApplicationStageName: string;
   public selectedApplicationRegionName: string;
 
-  abstract getCurrentStatus(): Observable<StatusReport>;
-
-  abstract alarm(): void;
-
-  protected extractDate(res: StatusReport): StatusReport {
-    res.timestamp = new Date(res.timestamp);
-    return res;
-  }
+  public selectedDomain: Domain;
 
   public selectApplicationInstance(appName: string, region: string, stage: string) {
+    console.log('Set Application Instance ', appName, region, stage);
     this.selectedApplicationName = appName;
     this.selectedApplicationStageName = stage;
     this.selectedApplicationRegionName = region;
   }
 
-  abstract getStages(): { name: string, stages: string[] }[];
+  getCurrentStatus(domain: Domain): Observable<StatusReport> {
+    return this.http.get<StatusReport>(domain.url).pipe(
+      map((res) => { res.timestamp = new Date(res.timestamp); return res; })
+    );
+  }
 
+  public selectDomain(domain: Domain) {
+    this.selectedDomain = domain;
+  }
 
+  public getSelectedDomain(): Domain {
+    return this.selectedDomain;
+  }
 
-
+  public getStages(): { name: string, stages: string[] }[] {
+      return this.selectedDomain.stages;
+  }
 }
