@@ -15,7 +15,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export const environment = {
-  production: true,
-  basePath: 'http://localhost:8080/kube/rest/'
-};
+export class Domain {
+
+  constructor(
+    public name: string,
+    public url: string,
+    public timestamp: number,
+    public stages: {
+      name: string,
+      stages: string[]
+    } []
+  ) {
+
+  }
+
+}
