@@ -21,5 +21,6 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  basePath: 'http://localhost:8080/kube/rest/'
 };
